requirements.txt
cachetools==3.0.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-vision==0.35.1
googleapis-common-protos==1.5.5
grpcio==1.16.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyMessager==1.0.1
python-dotenv==0.9.1
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
